OAuth2 commit 2 of probably 3
After much back and forth about how to implement OAuth, decided on a static "state" string for now. Because we are issuing cookies and because ASP.NET automatically issues an anti-forgery token, we're not that concerned right now about the XSS attempts, until we switch to JWT auth, then we can revisit the subject.

This commit also updates the UI, which had a bugfix pushed.